Change bot types

+++ botstruct.h
@@ -24,12 +24,12 @@ public:
 
     QList<int8_t> genome;
 
-    uint direction = 0;
+    uint8_t direction = 0;
     int energy;
     int minerals = 0;
     uint iterator = 0;
     uint old = 0;
-    uint type = ALIVE;
+    int8_t type = ALIVE;
 
     uint usedPhotosynthesis = 0, usedMinerals = 0, usedEat = 0;     // used gens
     uint mineralsCount = 0, photosynthesisCount = 0, eatCount = 0;  // gen count
